More tests for new Cata(). Tests extracted to a separate class. Simple class extracted to a separate file.
