+++ Composite/Composite.Cs.Tests/Simple.cs
@@ -1,0 +1,7 @@
+namespace Composite.Cs.Tests
+{
+    internal class Simple
+    {
+        public int Number { get; set; }
+    }
+}
